jetson/requirements.txt
numpy>=1.19.5
opencv-python>=4.5.3
sphero-sdk-raspberrypi-python>=0.1.1
pyserial>=3.5
smbus2>=0.4.1
filterpy>=1.4.5
scipy>=1.5.4
asyncio>=3.4.3
